# task-management/task-management-service/openapi.yaml
openapi: 3.0.0
info:
  title: Task Management API
  description: API for managing tasks.
  version: "1.0"
paths:
  /tasks:
    get:
      summary: List all tasks or filter by upcoming and/or email
      parameters:
        - in: query
          name: upcoming
          schema:
            type: string
          description: Filter by upcoming tasks
        - in: query
          name: email
          schema:
            type: string
          description: Filter by email address
      responses:
        '200':
          description: A list of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '201':
          description: The task was successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
  /tasks/{id}:
    put:
      summary: Update a task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
          description: Unique identifier of the task to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '200':
          description: The task was successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
    delete:
      summary: Delete a task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
          description: Unique identifier of the task to delete
      responses:
        '200':
          description: The task was successfully deleted
components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        title:
          type: string
        dueDate:
          type: string
          format: date-time
        email:
          type: string
          format: email
      required:
        - name
        - title
        - dueDate
        - email